Revamp of the loose files loading system.
You now place a file called loose.txt in the game's root directory. It
contains a list of directories to search with a directory per line. The
first listed directory has the highest priority, with the last having
the lowest priority.

It will always search the game's root directory last, even if it's not
in the list of directories. If you want to search the root directory at
higher priority use . as the directory name in the list.

If the file loose.txt doesn't exist then the game will act like the
vanilla game and not load loose files.